+++ CHANGELOG.md
@@ -49,6 +49,7 @@ Click up arrow to go back to TOC.
 - Add options to control every ffmpeg command. [issue #43](https://github.com/BingLingGroup/autosub/issues/43)
 - Add temp file save function. [issue #22](https://github.com/BingLingGroup/autosub/issues/22)
 - Add only audio fragments output. [issue #44](https://github.com/BingLingGroup/autosub/issues/44)
+- Add subtitles(first line) language detection powered by googletrans. [issue #40](https://github.com/BingLingGroup/autosub/issues/40)
 
 #### Changed(Unreleased)